requirements.txt
avro
confluent-kafka
PyYAML
